requirements.txt
amqp==2.4.2
billiard==3.5.0.5
canvasapi==0.13.0
celery==4.1.1
certifi==2018.11.29
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cx-Oracle==7.1.3
dj-database-url==0.5.0
Django==2.1.5
django-admin-autocomplete-filter==0.2
django-autocomplete-light==3.3.5
django-celery-beat==1.4.0
django-cors-headers==3.0.2
django-crispy-forms==1.7.2
django-extensions==2.1.6
django-filter==2.1.0
django-filters==0.2.1
django-heroku==0.3.1
django-rest-swagger==2.2.0
django-timezone-field==3.0
djangorestframework==3.9.0
djangorestframework-link-header-pagination==0.1.1
graphviz==0.10.1
gunicorn==19.9.0
idna==2.8
itypes==1.1.0
Jinja2==2.10
kombu==4.5.0
Markdown==3.1.1
MarkupSafe==1.1.0
openapi-codec==1.3.2
psycopg2==2.7.7
pydot==1.4.1
pyparsing==2.4.0
python-crontab==2.3.6
python-dateutil==2.8.0
pytz==2018.9
redis==3.2.1
requests==2.21.0
simplejson==3.16.0
six==1.12.0
uritemplate==3.0.0
urllib3==1.24.1
vine==1.3.0
whitenoise==4.1.2
